+++ lightcurve/pyproject.toml
@@ -12,7 +12,7 @@ packages = [
     { include = "object_api", from = "src" },
     { include = "crossmatch_api", from = "src" },
     { include = "probability_api", from = "src" },
-    { include = "database", from = "src" },
+    { include = "repository", from = "src/core" },
 ]
 
 [tool.poetry.dependencies]
