+++ lightcurve/src/lightcurve_api/api.py
@@ -6,8 +6,8 @@ from prometheus_fastapi_instrumentator import Instrumentator
 from ralidator_fastapi.ralidator_fastapi import RalidatorStarlette
 
 from core.config.config import app_config
-from database.sql import connect as connect_sql
-from database.sql import session_wrapper
+from core.repository.connection import connect as connect_sql
+from core.repository.connection import session_wrapper
 
 from .filters import get_filters_map
 from .routes import htmx, period, rest
